update user messages
